Rendering optimizations for infinite scroll comps
Each infinite scroll component had scrollTop and optionally scrollHeight
state variables. It was impacting performance badly when state was set on each scroll event. Instead the state variables were replaced with instance variables. Everything that is not used in render should not be part of a state. Additionally, for ChatPanel the ScrollBottomIcon state is handled in onScrollStop rather than onUpdate callback. It allows much smoother rendering of the icon.